feat: adds new error classes to llm and databases + introduces loglevel and logging from child error
