+++ cognee/infrastructure/databases/exceptions/__init__.py
@@ -8,4 +8,5 @@ from .exceptions import (
     EntityNotFoundError,
     EntityAlreadyExistsError,
     DatabaseNotCreatedError,
+    EmbeddingException,
 )
