+++ cognee/infrastructure/databases/vector/exceptions/exceptions.py
@@ -16,5 +16,7 @@ class CollectionNotFoundError(CogneeValidationError):
         message,
         name: str = "CollectionNotFoundError",
         status_code: int = status.HTTP_422_UNPROCESSABLE_ENTITY,
+        log=True,
+        log_level="DEBUG",
     ):
-        super().__init__(message, name, status_code)
+        super().__init__(message, name, status_code, log, log_level)
